ARM: dts: Set the LED default-state to "off"
The default-state property for an LED determines the state of the LED
when its trigger is set to "none". Many users expect that specifying a
trigger of "none" means that the LED will be off, and there is a logic
to that assumption, but the current default of "keep" makes the
assumption false.

Given that no Pi models have a default-trigger of "none", changing the
default-state should have no visible impact for most people, but doing
so will permit a more obvious and more consistent way of turning off
the LEDs.

Note that 6.2 included an upstream DT reorganisation that required some
downstream changes, and while doing that I changed the default states.
This commit brings rpi-6.1.y into line with rpi-6.2.y.

See: https://github.com/raspberrypi/firmware/issues/1742

Signed-off-by: Phil Elwell <phil@raspberrypi.com>